Moved the repositories to their own folder
